--warning
git-svn-id: svn+ssh://forge.imag.fr/var/lib/gforge/chroot/scmrepos/svn/life/trunk/life/trunk@5577 18f2cc81-8059-4896-b63e-f2d89ec8fd72

+++ life/lifediscr/discontinuousinterfaces.hpp
@@ -217,7 +217,7 @@ public:
             boost::tie( fit, fen ) = M_mesh.elementsRange();
             while( fit != fen )
                 {
-                    if ( fit->marker().value() == boost::get<1>( marker ) )
+                    if ( fit->marker().value() == (int)boost::get<1>( marker ) )
                         {
                             M_dof.addDofFromElement( *fit, next_free_dof, 0 );
                         }
